refactor: move select-travel-token-screen into screen-components
