+++ src/stacks-hierarchy/Root_SelectTravelTokenScreen.tsx
@@ -1,4 +1,4 @@
-import {SelectTravelTokenScreenComponent} from '@atb/select-travel-token-screen';
+import {SelectTravelTokenScreenComponent} from '@atb/screen-components/select-travel-token-screen';
 import {RootStackScreenProps} from './navigation-types';
 
 type Props = RootStackScreenProps<'Root_SelectTravelTokenScreen'>;
