+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_TravelTokenScreen/Profile_TravelTokenScreen.tsx
@@ -10,7 +10,7 @@ import {FaqSection} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_Pr
 import {useTokenToggleDetailsQuery} from '@atb/mobile-token/use-token-toggle-details';
 import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
 import {Section, GenericSectionItem} from '@atb/components/sections';
-import {TokenToggleInfo} from '@atb/select-travel-token-screen';
+import {TokenToggleInfo} from '@atb/screen-components/select-travel-token-screen';
 
 export const Profile_TravelTokenScreen = () => {
   const styles = useStyles();
